+++ changelog.md
@@ -1,9 +1,15 @@
-## 0.0.2
+## [0.0.2](https://github.com/Ackimixs/anOtherDiscordPlayer/commit/4b194ec41649da4cb3dad65c0bc6a9ee75e57270)
  - initial commit
  - basic bot functionality focused on music (and openAi)
  - added twitch, YouTube and spotify support
 
-## 0.1.0
+## [0.1.0](https://github.com/Ackimixs/anOtherDiscordPlayer/commit/d817a8674c57d670e7bd1c6ce06c0a7cdbdb2c0b)
  - add autocomplete support for every search command (youtube, spotify, twitch)
  - patch player and queue class (all that bugs was find during my math classes)
  - update many things
+
+## [0.2.0](https://github.com/Ackimixs/anOtherDiscordPlayer/commit/41a6b363c4959289a033d1a5f714f2e178a632ce) 
+- use of new youtube api, instead of the official i use [youtube-sr](https://www.npmjs.com/package/youtube-sr)  
+ => to get more easy access to the data
+- add new different command for music
+- update many other things